requirements.txt
fastapi
python-multipart
uvicorn[standard]
jinja2
sudachipy
sudachidict-small

python-dotenv